+++ packages/multichain-account-service/src/types.ts
@@ -1,6 +1,8 @@
 import type {
   Bip44Account,
   MultichainAccountGroup,
+  MultichainAccountWalletId,
+  MultichainAccountWalletStatus,
 } from '@metamask/account-api';
 import type {
   AccountsControllerAccountAddedEvent,
@@ -74,11 +76,6 @@ export type MultichainAccountServiceAlignWalletsAction = {
   handler: MultichainAccountService['alignWallets'];
 };
 
-export type MultichainAccountServiceGetIsAlignmentInProgressAction = {
-  type: `${typeof serviceName}:getIsAlignmentInProgress`;
-  handler: MultichainAccountService['getIsAlignmentInProgress'];
-};
-
 export type MultichainAccountServiceCreateMultichainAccountWalletAction = {
   type: `${typeof serviceName}:createMultichainAccountWallet`;
   handler: MultichainAccountService['createMultichainAccountWallet'];
@@ -98,7 +95,6 @@ export type MultichainAccountServiceActions =
   | MultichainAccountServiceSetBasicFunctionalityAction
   | MultichainAccountServiceAlignWalletAction
   | MultichainAccountServiceAlignWalletsAction
-  | MultichainAccountServiceGetIsAlignmentInProgressAction
   | MultichainAccountServiceCreateMultichainAccountWalletAction;
 
 export type MultichainAccountServiceMultichainAccountGroupCreatedEvent = {
@@ -111,13 +107,19 @@ export type MultichainAccountServiceMultichainAccountGroupUpdatedEvent = {
   payload: [MultichainAccountGroup<Bip44Account<KeyringAccount>>];
 };
 
+export type MultichainAccountServiceWalletStatusChangeEvent = {
+  type: `${typeof serviceName}:walletStatusChange`;
+  payload: [MultichainAccountWalletId, MultichainAccountWalletStatus];
+};
+
 /**
  * All events that {@link MultichainAccountService} publishes so that other modules
  * can subscribe to them.
  */
 export type MultichainAccountServiceEvents =
   | MultichainAccountServiceMultichainAccountGroupCreatedEvent
-  | MultichainAccountServiceMultichainAccountGroupUpdatedEvent;
+  | MultichainAccountServiceMultichainAccountGroupUpdatedEvent
+  | MultichainAccountServiceWalletStatusChangeEvent;
 
 /**
  * All actions registered by other modules that {@link MultichainAccountService}